requirements.txt
discord.py==2.1.0
httpx==0.23.1
python-dotenv==0.21.0